// src/types/dto/SiteDto.ts
export interface SiteDto {
  id: string;
  ownerId: string;
  siteName: string;
  defaultDisplayTitle: boolean;
  defaultDisplayLogo: boolean;
  defaultTitleFont: string;
  defaultTitlePosition: string;
  defaultLogoPosition: string;
  defaultLogoUrl: string;
  defaultFontWeight: string;
  defaultFontColor: string;
  defaultFontSize: number;
}